Library: the audit's remaining bugs in edit, lists, files and devices
- Notepad with Word Wrap: two carets blinked where a line wraps, one at
  the end of the row and one at the start of the next. Only the lower
  row draws it now. A long unbroken Cyrillic or CJK word was cut in the
  middle of a letter, both halves drawn as boxes; the break falls
  between characters.
- An input method's commit longer than the key buffer was dropped whole,
  nothing typed. It is fetched again with room enough and inserted.
- A quick click on row 0 of one list and then row 0 of another opened
  the second one's item as a double-click; the same across two edit
  boxes selected a word. The double-click is per control now.
- Page Up and Page Down in Large Icons moved one row's worth of items,
  and in List a single item. They move a screenful.
- Paste after the Snipping Tool had copied a picture froze the program
  for 0.4 s and gave nothing: it asked itself and could not answer. It
  knows it holds a picture, not text.
- A PNG whose last block failed to write (disk full) was reported saved.
- Properties: Apply and then Cancel had renamed or changed the item, but
  Explorer and the desktop showed the old name until F5; they refresh.
  After Apply on a symbolic link the permission boxes all turned on (the
  link's own 0777) instead of the target's.
- Recycle Bin: a file from another drive whose name was already in the
  bin as a folder could not be recycled; the check looked at the bin's
  entry, not the file. Copies across drives lost their mode, so a script
  came back not executable and a private file readable by all. A
  restore whose bin copy could not be removed left it there without a
  record; it backs out now.
- My Computer could leave out a drive at /mnt/<one letter>, read past
  its name by the /media/<user> test.
- Check Now and the label change: closing the wait box early left the
  program frozen in wait, a chatty e2fsck blocked on a full pipe. The box
  stays until the job ends, and an early stop reads the output through.
- Device Manager: a battery read 'Battery — 85Discharging' and is now
  'Battery — 85%, Discharging'; a PCI device with no driver showed
  'Driver: (null)' and Driver Details looked up a module called (null).
- A .ts file opened in VLC while Explorer called it a source file: a
  TypeScript file is text; only an MPEG stream (its 0x47 sync bytes)
  goes to the video player.
- Explorer's folder default was written as Exec=<path> %f unquoted: a
  path with a space left xdg-open with no program for folders.
- 32-bit builds: battery percentage and time left wrapped negative for a
  pack over about 21 Wh; an icon with a huge colour count could slip
  past its size check. MonitorOff above 18 hours wrapped DPMS to
  minutes; it is capped at what the protocol holds. An Aero corner
  sheet narrower than its cells was read past its end.

+++ lib/power.c
@@ -80,7 +80,9 @@ int w2k_power_read(W2kPower *p)
                 full = read_num(dir, "charge_full", -1);
                 rate = read_num(dir, "current_now", -1);
             }
-            if (pct < 0 && now >= 0 && full > 0) pct = now * 100 / full;
+            /* In 64 bits: a 50 Wh pack reads 50,000,000, and times 100 or
+             * 60 that wrapped negative where long is 32 bits (i386, armhf). */
+            if (pct < 0 && now >= 0 && full > 0) pct = (long)((long long)now * 100 / full);
             if (pct < 0) continue;                 /* nothing to show */
             if (pct > 100) pct = 100;
             p->present = 1;
@@ -96,7 +98,7 @@ int w2k_power_read(W2kPower *p)
             if (rate > 0 && now >= 0 && full > 0) {
                 long left = p->charging == 1 ? full - now : now;
                 if (left < 0) left = 0;
-                p->minutes_left = (int)(left * 60 / rate);
+                p->minutes_left = (int)((long long)left * 60 / rate);
             }
         } else if (!strcmp(type, "Mains") || !strcmp(type, "USB") ||
                    !strncmp(type, "ADP", 3)) {
